Various fixes for win32.

+++ src/client/cl_scrn.c
@@ -642,11 +642,11 @@ int entitycmpfnc( const entity_t *a, const entity_t *b )
 	*/
 	if ( a->model == b->model )
 	{
-		return ( ( int ) a->skin - ( int ) b->skin );
+		return (int)( (size_t)a->skin - (size_t)b->skin );
 	}
 	else
 	{
-		return ( ( int ) a->model - ( int ) b->model );
+		return (int)( (size_t)a->model - (size_t)b->model );
 	}
 }
